chore: combine both universal auth code flow suites
