+++ test/integration/login/code-flow.spec.ts
@@ -2,8 +2,23 @@ import { describe, it, expect } from "vitest";
 import { getEnv } from "../helpers/test-client";
 import { oauthApp } from "../../../src/app";
 import { testClient } from "hono/testing";
+import { EmailOptions } from "../../../src/services/email/EmailOptions";
+import { snapshotResponse } from "../helpers/playwrightSnapshots";
+import { FOKUS_VENDOR_SETTINGS } from "../../fixtures/vendorSettings";
+import { AuthorizationResponseType } from "../../../src/types";
 
-describe("Login with code", () => {
+function getCodeAndTo(email: EmailOptions) {
+  const codeEmailBody = email.content[0].value;
+  // this gets the space before so we don't match CSS colours
+  const codes = codeEmailBody.match(/(?!#).[0-9]{6}/g)!;
+  const code = codes[0].slice(1);
+
+  const to = email.to[0].email;
+
+  return { code, to };
+}
+
+describe("Universal Auth login with code", () => {
   it("should return a 400 if there's no code", async () => {
     const env = await getEnv();
     const oauthClient = testClient(oauthApp, env);
@@ -18,5 +33,194 @@ describe("Login with code", () => {
     });
 
     expect(incorrectCodeResponse.status).toBe(400);
+  });
+  it("should login with code", async () => {
+    const env = await getEnv({
+      vendorSettings: FOKUS_VENDOR_SETTINGS,
+      testTenantLanguage: "nb",
+    });
+    const oauthClient = testClient(oauthApp, env);
+
+    const response = await oauthClient.authorize.$get({
+      query: {
+        client_id: "clientId",
+        response_type: AuthorizationResponseType.TOKEN_ID_TOKEN,
+        scope: "openid",
+        redirect_uri: "http://localhost:3000/callback",
+        state: "state",
+      },
+    });
+
+    expect(response.status).toBe(302);
+    const location = response.headers.get("location");
+
+    expect(location!.startsWith("/u/code")).toBeTruthy();
+
+    const stateParam = new URLSearchParams(location!.split("?")[1]);
+
+    const query = Object.fromEntries(stateParam.entries());
+
+    const codeInputFormResponse = await oauthClient.u.code.$get({
+      query: {
+        state: query.state,
+      },
+    });
+
+    expect(codeInputFormResponse.status).toBe(200);
+
+    await snapshotResponse(codeInputFormResponse);
+
+    const postSendCodeResponse = await oauthClient.u.code.$post({
+      query: { state: query.state },
+      form: {
+        username: "foo@example.com",
+      },
+    });
+
+    expect(postSendCodeResponse.status).toBe(302);
+    const enterCodeLocation = postSendCodeResponse.headers.get("location");
+
+    const { to, code } = getCodeAndTo(env.data.emails[0]);
+    expect(to).toBe("foo@example.com");
+
+    // Authenticate using the code
+    const enterCodeParams = enterCodeLocation!.split("?")[1];
+    const enterCodeQuery = Object.fromEntries(
+      new URLSearchParams(enterCodeParams).entries(),
+    );
+
+    const enterCodeForm = await oauthClient.u["enter-code"].$get({
+      query: { state: enterCodeQuery.state },
+    });
+    expect(enterCodeForm.status).toBe(200);
+    await snapshotResponse(enterCodeForm);
+
+    const authenticateResponse = await oauthClient.u["enter-code"].$post({
+      query: {
+        state: enterCodeQuery.state,
+      },
+      form: {
+        code,
+      },
+    });
+
+    const codeLoginRedirectUri = authenticateResponse.headers.get("location");
+    const redirectUrl = new URL(codeLoginRedirectUri!);
+    expect(redirectUrl.pathname).toBe("/callback");
+    const hash = new URLSearchParams(redirectUrl.hash.slice(1));
+    const accessToken = hash.get("access_token");
+    expect(accessToken).toBeTruthy();
+    const idToken = hash.get("id_token");
+    expect(idToken).toBeTruthy();
+  });
+
+  it('snapshot desktop "enter code" form', async () => {
+    const env = await getEnv({
+      vendorSettings: FOKUS_VENDOR_SETTINGS,
+      testTenantLanguage: "nb",
+    });
+    const oauthClient = testClient(oauthApp, env);
+
+    const response = await oauthClient.authorize.$get({
+      query: {
+        client_id: "clientId",
+        response_type: AuthorizationResponseType.TOKEN_ID_TOKEN,
+        scope: "openid",
+        redirect_uri: "http://localhost:3000/callback",
+        state: "state",
+      },
+    });
+    const location = response.headers.get("location");
+    const stateParam = new URLSearchParams(location!.split("?")[1]);
+
+    const query = Object.fromEntries(stateParam.entries());
+
+    const codeInputFormResponse = await oauthClient.u.code.$get({
+      query: {
+        state: query.state,
+      },
+    });
+
+    expect(codeInputFormResponse.status).toBe(200);
+
+    await snapshotResponse(codeInputFormResponse, "lg");
+  });
+
+  it('snapshot mobile "enter code" form', async () => {
+    const env = await getEnv({
+      vendorSettings: FOKUS_VENDOR_SETTINGS,
+      testTenantLanguage: "nb",
+    });
+    const oauthClient = testClient(oauthApp, env);
+
+    const response = await oauthClient.authorize.$get({
+      query: {
+        client_id: "clientId",
+        response_type: AuthorizationResponseType.TOKEN_ID_TOKEN,
+        scope: "openid",
+        redirect_uri: "http://localhost:3000/callback",
+        state: "state",
+      },
+    });
+    const location = response.headers.get("location");
+    const stateParam = new URLSearchParams(location!.split("?")[1]);
+
+    const query = Object.fromEntries(stateParam.entries());
+
+    const codeInputFormResponse = await oauthClient.u.code.$get({
+      query: {
+        state: query.state,
+      },
+    });
+
+    expect(codeInputFormResponse.status).toBe(200);
+
+    await snapshotResponse(codeInputFormResponse, "sm");
+  });
+
+  it("should reject bad code", async () => {
+    const env = await getEnv();
+    const oauthClient = testClient(oauthApp, env);
+
+    const response = await oauthClient.authorize.$get({
+      query: {
+        client_id: "clientId",
+        response_type: AuthorizationResponseType.TOKEN_ID_TOKEN,
+        scope: "openid",
+        redirect_uri: "http://localhost:3000/callback",
+        state: "state",
+      },
+    });
+
+    const location = response.headers.get("location");
+
+    const stateParam = new URLSearchParams(location!.split("?")[1]);
+
+    const query = Object.fromEntries(stateParam.entries());
+
+    const postSendCodeResponse = await oauthClient.u.code.$post({
+      query: { state: query.state },
+      form: {
+        username: "foo@example.com",
+      },
+    });
+    const enterCodeLocation = postSendCodeResponse.headers.get("location");
+
+    const enterCodeParams = enterCodeLocation!.split("?")[1];
+    const enterCodeQuery = Object.fromEntries(
+      new URLSearchParams(enterCodeParams).entries(),
+    );
+
+    const incorrectCodeResponse = await oauthClient.u["enter-code"].$post({
+      query: {
+        state: enterCodeQuery.state,
+      },
+      form: {
+        // clearly wrong!
+        code: "123456",
+      },
+    });
+
+    await snapshotResponse(incorrectCodeResponse);
   });
 });
